Fixed a lot of things, mainly making stats more consistent over different part of the program, and making CNA and clonality analysis more robust.
Now identifies germline variants in the clonality analysis (variants sating constant with 100% clonality for all samples). Variants that are filtered out in the clonal analysis are now added back in after the clonal stories are determined, merging them with the closest fit if the filtered story is consistent with the clone story.